added authentication to whole app

+++ frontend/src/components/workoutDetails.js
@@ -1,14 +1,23 @@
 import { useWorkoutsContext } from "../Hooks/useWorkoutsContext";
+import { useAuthContext } from "../Hooks/useAuthContext";
 
 //date fns
 import formatDistanceToNow from "date-fns/formatDistanceToNow";
 
 const WorkOutDetails = ({ workout }) => {
   const { dispatch } = useWorkoutsContext();
+  const { user } = useAuthContext();
 
   const handleClick = async () => {
+    if (!user) {
+      return;
+    }
+
     const response = await fetch("/api/workouts/" + workout._id, {
       method: "DELETE",
+      headers: {
+        Authorization: `Bearer ${user.token}`,
+      },
     });
     const json = await response.json();
 
@@ -25,7 +34,7 @@ const WorkOutDetails = ({ workout }) => {
         {workout.load}
       </p>
       <p>
-        <strong>Reps : (kg)</strong>
+        <strong>Reps: </strong>
         {workout.reps}
       </p>
       <p>
